pyproject.toml
[project]
name = "langchain-project"
version = "0.1.0"
description = "A LangChain project"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "anthropic>=0.18.0",
    "langfuse>=2.0.0",
    "dotaconstants>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "black>=24.0.0",
]

[tool.poetry]
name = "langchain-project"
version = "0.1.0"
description = "A LangChain project"
authors = ["Developer"]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.3.0"
langchain-core = "^0.3.0"
langchain-community = "^0.3.0"
langchain-openai = "^0.2.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
anthropic = "^0.18.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.0.0"
ruff = "^0.4.0"
mypy = "^1.8.0"
black = "^24.0.0"

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
